add(feat): display markers and manage them
This commit adds the updated MapScreen.kt which include
AnnotationManager that handles the markers on the map. There is also the
pipelining of data from the viewModel but this isn't working correctly
for now. The zoom is clamped. Also add the test suite for the
MapScreen.kt which test that the zoom is really clamped. Finally add
ressources which is the .png file for the red marker. Also update the
zoomControls for better testing.