Feature implement interaction with markers (#99)
* add(feat) : make the markers clickable

This commit define the functionality that makes the markers clickable. On click the camera centers on the corresponding marker and display a preview of the cardScreen. The handling of the different listeners and viewAnnotation needs to be fixed.

* fix: switch between markers

The switch between markers make the transition from one preview to another. Before, it just displayed the clicked marker preview wihtout deleting the previous preview. Need to fix the collisation between the CameraChangeListener and the MapIdleListener.

* fix: collision between listeners

Fix the previous problem about the collision of the MapIdleListener and CameraChangeListener for the preview of the markers by adding a boolean.

* add: add documentation

* fix: take PR reviews into account

Remove unnecessary elements highlighted in the reviews

* fix: format MapScreen using ktfmt format

---------

Co-authored-by: Perazzolo Sean <47214991+sean-prz@users.noreply.github.com>