feat: link the markers to the parking view model
This commit add the new feature that enables us to navigate to the selected parking if we click on the preview card and the preview card display relevant info about the selected parking. parking is binded to the markers as a JSon object and we deserialize it to get all the data of the marker.