Link the markers to the parking view model (#114)
* feat: link the markers to the parking view model

This commit add the new feature that enables us to navigate to the selected parking if we click on the preview card and the preview card display relevant info about the selected parking. parking is binded to the markers as a JSon object and we deserialize it to get all the data of the marker.

* fix(test) : use mock repo for the MapScreen Test

This commit implements the offline mock repos (ParkingRepo and ImageRepo) so that future test will be more efficient.

* format : format MapScreenTest.kt using ktfmt format

* fix(map) : fix issue when reloading the map

When reloading the map the markers are not displayed unless we dezoom. This commmit fix that by updated the LaunchedEffect of MapScreen.kt

* fix : take into account reviews from PR

Fixes variable name parkingDeserialized to fit camelCase. pass cancelables from mutable list of Cancelable to Cancelable since there is only one Cancelable. Refactor gson outside the Scaffold to avoid recreating it every time