BugFix Preview Card don't display after adding a review (#134)
* fix: fix preview card bug

The preview card doesn't display after adding a review. The fix was to refactor the MapIdleCallBack listener to a var instead of a val inside the clickListener

* fix: fix bad management of listener

The fix convert listener var to a remember and remove the null attribution

* fix: bug updating the map

The map update when a parking is added and not when an existing one is updated. I updated the LaunchedEffect to look for the number of review on selected parking since it is the onyl field that can be modified