refactor: easier implementation for getParkingsInRadius
This commit simplifies the function.
Instead of getting all tiles that overlaps the circle and calling getParkingsInRect on each tile, it just calls getParkingsInRect once with as argument
the smallest rectangle that contains the circle.

The test have been adjusted as well as one function of the Tile class.
In paticular I had to revert the changes to the testInstances to pass with the test of the Profile screen.