Theme (#199)
* refactor: remove alpha copies of colors in theme

Add a debug theme for easy identification of colors.
Removed darkColorScheme (now same as light color scheme)

* fix: harmonized theme app-wise

* feature: redefine light theme colors

* test: comment test for dark theme

As dark theme is the same as light theme at this point.

* fix: minute details in theme

- Google SignIn button has black text
- Anonymous SignIn button is now in primary color level
- FilterSection has white background
- ToggleButton have blue text when off
- Labels text of input are now black

* doc: rectify color selectors documentation

Was a bad copy-past before.

* test: remove FilterSection preview

* test: remove debug theme

As it is no longer useful.

* feature: apply theme to AllReviewsScreen filters

* test: update unit test for ThemeTest

* feature: redo logo and its integration

Our app logo is now showed during the launch of our application. It is also used by default for notification and multiple other locations like Toast or even the Google Play Store if it ever reach this state.

* refactor: remove ununsed alpha modifier in theme

* refactor: text are now mostly onBackground

Add a getTextColorField to better link the colors in the different atoms.

* fix: make clearInputText button visible

Change its color from onPrimary to defaultOnColor.

* fix: make parking rectangle red when too big

* test: revise an InputText call with new arguments

* test: revise an InputText call with new arguments

* feature: colors for checkbox elements

* refactor: cleaned the Color class

Removed ununsed import and colors.