Search Bar and Geocoding (#309)
* add(UI) : rework MapScreen to integrate Search Bar

* feat : add the backend for geocoding.

This commit add the backend (geocoding) not yet functional.
 The bug comes from the deserialization of the location. This also adds the bottom screen that is displayed with the suggestions.

* bugfix : correct the deserialization error

This commit address the error with the deserialization of the location. Now the feature work as intended.

* add: padding correction for suggestion

* refactor: format and refactor strings to strings.xml

Formatting was done using ktfmt format and refactor of all the strings into strings.xml

* refactor: refactor color and theme and correction

The correction is about a rework of the UI of the clear button of the Search Bar. This commit refactors all the content of the feature so that it uses the color and theme of the package theme

* add: test suite for the new feature

* add: test suite for addressviewModel and update tests

* refactor: update tests in MainActivityTest and MapScreenTest

* format : format using ktfmf format

* fix: correct tests and remove unecessary comment

* refactor: increase modularity in MapScreen

* add : add missing documentation