add: SuggestionMenu and search bar behavior
The suggestion menu is now placed in the screen. The Search Bar has the expected behavior now. List Screen suggestion Menu is limited to 7 suggestion (exxcluding my location). Still needs to work on the formatting of the suggestions, and the policy to manage nominatim answer. The chosen location isn't updating the list of parkings