Feature Search Bar for List Screen (#342)
* feat: search feature for list Screen

Rework the list screen UI to integrate search bar and suggestion. Add partially complete logic for the feature.

* add: SuggestionMenu and search bar behavior

The suggestion menu is now placed in the screen. The Search Bar has the expected behavior now. List Screen suggestion Menu is limited to 7 suggestion (exxcluding my location). Still needs to work on the formatting of the suggestions, and the policy to manage nominatim answer. The chosen location isn't updating the list of parkings

* fix: fix formatting problem for suggestions

* add : logic updates and design ajustement

Formatting of the suggestion can be longer for the list screen and the cyrcle center is updated when the user click on a suggestion

* refactor: refactor the Search bar

* refactor: refactor and rebase the code

* fix: fix Louis comment

* fix : update suggestion format and add documentation

Suggestion format for the map integrates region field.

* fix : update and add tests suites

* add : enum for the mode of formatting

* fix : fix maria comments

* fix : fix maria comments bis

* fix : update tests for address repository

* rebase

* fix: padding search bar list screen